src/cli/requirements.txt
# Core CLI dependencies
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0

# HTTP requests
requests>=2.28.0

# HuggingFace Hub for model downloads
huggingface-hub>=0.16.0

# Machine Learning dependencies
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Document processing
pypdf2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0

# Configuration and data handling
toml>=0.10.2
jsonschema>=4.17.0

# Progress bars and UI
tqdm>=4.64.0

# Logging
colorlog>=6.7.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
